+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Use [Jim Tcl]
 - Do not reset the environment variable `IAMROOT_LIB` to the iamroot library
 - `ish(1)` is now using `ld-iamroot.so(8)`
+- Do not forward the none-`at` functions `access()` and `euidaccess()` to the
+  `at` functions `faccessat()`
 
 ## [v23] - 2024-06-25